[FRONTEND] Move table components to other folder
